pyproject.toml
[tool.poetry]
name = "vitaliy-avdoshkin-coursework-3"
version = "0.1.0"
description = ""
authors = ["Vitaliy.Avdoshkin <wioswa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas-stubs = "^2.2.2.240603"
types-requests = "^2.32.0.20240712"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
#line-length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '''
/(
  | \.venv
)/
'''



